+++ packages/state-transition/src/block/processDeposit.ts
@@ -15,7 +15,7 @@ import {
 import {DepositData} from "@lodestar/types/lib/phase0/types.js";
 import {DepositReceipt} from "@lodestar/types/lib/electra/types.js";
 import {ZERO_HASH} from "../constants/index.js";
-import {computeDomain, computeSigningRoot, increaseBalance} from "../util/index.js";
+import {computeDomain, computeSigningRoot, hasCompoundingWithdrawalCredential, hasEth1WithdrawalCredential, increaseBalance, switchToCompoundingValidator} from "../util/index.js";
 import {CachedBeaconStateAllForks, CachedBeaconStateAltair, CachedBeaconStateElectra} from "../types.js";
 
 /**
@@ -90,6 +90,13 @@ export function applyDeposit(
         amount: BigInt(amount),
       });
       stateElectra.pendingBalanceDeposits.push(pendingBalanceDeposit);
+
+      if (hasCompoundingWithdrawalCredential(withdrawalCredentials) 
+        && hasEth1WithdrawalCredential(validators.getReadonly(cachedIndex).withdrawalCredentials)
+        // TODO Electra: is_valid_deposit_signature
+      ) {
+        switchToCompoundingValidator(stateElectra, cachedIndex);
+      }
     }
   }
 }
