+++ packages/state-transition/src/util/index.ts
@@ -24,3 +24,4 @@ export * from "./syncCommittee.js";
 export * from "./validator.js";
 export * from "./weakSubjectivity.js";
 export * from "./deposit.js";
+export * from "./electra.js";
